Move MultiStreamableMessageSource to the right module
The MultiStreamableMessageSource resides in the event sourcing module,
while it belongs in the messaging module and event handling package. We
should thus copy the implementation and extend the old one from the copy
Doing so ensure we don't break users right away. The eventsourcing
variant should be deprecated as well.

#bug/multi-source-tracking-token
